requirements.txt
gradio==3.4
librosa>=0.9.2
matplotlib==3.5.3
numpy>=1.21.6

pycwt>=0.3.0a22
pypinyin>=0.47.1
PyYAML>=6.0
scipy>=1.7.3
six>=1.16.0
soundfile>=0.11.0
tqdm>=4.64.1
ipywidgets>=7.6.5
